refactoring hyperopt into its own package
