+++ ludwig/utils/misc.py
@@ -18,6 +18,7 @@ import copy
 import os
 import random
 import subprocess
+import subprocess as sp
 import sys
 from collections import OrderedDict
 from collections.abc import Mapping
@@ -160,3 +161,18 @@ def get_class_attributes(c):
         i for i in dir(c)
         if not callable(getattr(c, i)) and not i.startswith("_")
     )
+
+
+def get_available_gpu_memory():
+    _output_to_list = lambda x: x.decode('ascii').split('\n')[:-1]
+
+    COMMAND = "nvidia-smi --query-gpu=memory.free --format=csv"
+    try:
+        memory_free_info = _output_to_list(sp.check_output(COMMAND.split()))[
+                           1:]
+        memory_free_values = [int(x.split()[0])
+                              for i, x in enumerate(memory_free_info)]
+    except Exception as e:
+        print('"nvidia-smi" is probably not installed.', e)
+
+    return memory_free_values
